% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGrid.R
\name{addGrid}
\alias{addGrid}
\title{Add a Grid to a Plot}
\usage{
addGrid(
  at_x,
  at_y,
  col = NA,
  border = "black",
  lwd = 1,
  lty = 1,
  box = FALSE,
  ...
)
}
\arguments{
\item{at_x}{Coordinates on the x-Axis where to draw vertical lines.}

\item{at_y}{Coordinates on the y-Axis where to draw horizontal lines.}

\item{col}{The color of the background.}

\item{border}{The color of lines (and box).}

\item{lwd}{The width of lines (see \code{par()}).}

\item{lty}{The type of lines (see \code{par()}).}

\item{box}{A boolean. If \code{TRUE} add a box around the plot area.}

\item{...}{Other graphical parameters as in \code{par()}.}
}
\description{
This function adds a grid to an existing plot (like \code{box()}). User can
specify coordinates on x- and/or y-axis where to draw vertical and/or
horizontal lines respectively. User can also color the background of the plot
area and add a box around this area (if required).
}
\details{
If user does not specify \code{at_x} and \code{at_y}, the grid is aligned with the tick
marks on the corresponding default axes (computed by \code{axTicks}).
}
\examples{
maps::map()
addGrid()
addGrid(box = TRUE)
addGrid(col = "#ff000044", border = "green", box = TRUE)

maps::map()
addGrid(at_x = axTicks(1), border = "blue", lwd = 2)
addGrid(at_y = axTicks(2), border = "red")
addGrid(at_x = NULL, at_y = NULL, lwd = 3, box = TRUE)
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@gmail.com}
}
\keyword{background}
\keyword{box,}
\keyword{grid,}
