% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixColors.R
\name{blendColors}
\alias{blendColors}
\title{Blend colors}
\usage{
blendColors(cols)
}
\arguments{
\item{cols}{set of colors to be mixed.}
}
\description{
Blend a set of colors and return the color thereby obtained.
}
\examples{
blendColors(c("blue", "purple"))
}
