% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box2.R
\name{box2}
\alias{box2}
\title{Box version 2}
\usage{
box2(bty = 1:4, col2fill = NULL, ...)
}
\arguments{
\item{bty}{A numerical/character vector or a character string decribing on which side the box is drawn (see details).}

\item{...}{Further graphical parameters (see ‘par’) may also be supplied as arguments, particularly, line type, ‘lty’, line width, ‘lwd’, color, ‘col’ and for ‘type = "b"’, ‘pch’.  Also the line characteristics ‘lend’, ‘ljoin’ and ‘lmitre’.}
}
\description{
Draw a Box around a Plot
}
\details{
This function is made to give more flexibility to the \code{\link[grpahics]{box}} function.
Users provides an object first coerced by ‘as.character’ to a character string that is secondly split
into single characters. For all of these characters, matches are sought with element among \code{c('1', '2', '3', '4', 'b', 'l', 't', 'r'}. All of these are matched with the
}
\examples{
# Example 1:
plot0()
box2()
##
plot0()
box2("14", col2fill="grey80", lwd=2)
plot0()
box2(c(1,4), col2fill="grey80", lwd=2)
plot0()
box2(c(1,4), col2fill="grey80", lwd=2)
# Example 2:
plot0(c(-10,10), asp=1)
}
\seealso{
\code{\link[graphics]{box}}
}
\keyword{box}

