% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/circle.R
\name{circle}
\alias{circle}
\title{Draw circles}
\usage{
circle(x = 0, y = x, radi = 1, from = 0, to = 2 * pi, incr = 0.01,
  closing = 1, col = 1, ...)
}
\arguments{
\item{x}{The x coordinates of the centers of the circles.}

\item{y}{The y coordinates of the centers of the circles.}

\item{radi}{The radii of the circles.}

\item{from}{Start points expressed in radians.}

\item{to}{End points expressed in radians.}

\item{incr}{Increments expressed in radians.}

\item{col}{Vector of colors of the circles.}

\item{...}{Additional arguments to be passed to either \code{\link{polygon}} function (if \code{boder} is defined) either \code{\link{lines}} function (otherwise).}
}
\description{
\code{circle} Draw circles in a flexible way.
}
\details{
The number of cicrles drawn is given by the maximum argument length among \code{x}, \code{y} and \code{radi} arguments. Sizes are adjusted using \code{\link{rep_len}} function. The plot used \code{\link{lines}} function, however if \code{border} is defined, then \code{\link{polygon}} function is used.
}
\examples{
#Example 1:
plot0()
circle()

#Example 2:
plot0(x=c(-2,2),y=c(-2,2))
circle(c(-1,1), col=2, lwd=4)

#Example 3:
plot0(x=c(-2,2),y=c(-2,2), asp=1)
circle(x=c(-1,1),c(1,1,-1,-1),from=pi*seq(0.25,1,by=0.25),to=1.25*pi, col=2, border=4, lwd=3)
}
\keyword{circle,}
\keyword{plot}

