% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circles.R
\name{circles}
\alias{circles}
\title{Draw circles}
\usage{
circles(x, y = x, radi = 1, from = 0, to = 2 * pi, incr = 0.01,
  pie = FALSE, add = TRUE, ...)
}
\arguments{
\item{x}{the x coordinates of the centers of the circles.}

\item{y}{the y coordinates of the centers of the circles.}

\item{radi}{the radii of the circles.}

\item{from}{the angles, expressed in radians, from which circles are drawn.}

\item{to}{the angles, expressed in radians, to which circles are drawn.}

\item{incr}{increments between two points to be linked (expressed in radians).}

\item{pie}{a logical. If \code{TRUE}, end points are linked with the center of the circle (default is set to \code{FALSE}).}

\item{add}{a logical. Should the circles be added on the current plot?}

\item{...}{additional arguments to be passed to \code{\link[graphics]{polygon}} function.}
}
\value{
An invisible list of \code{data.frame} of two columns including the
coordinates of all circles.
}
\description{
Draw circles in a flexible way.
}
\details{
The number of circles drawn is given by the maximum argument length amng
\code{x}, \code{y}, \code{radi}, \code{from} and \code{to} arguments.
Sizes are adjusted (i.e. repeated over) with \code{\link{rep_len}} function.
}
\examples{
# Example 1:
plot0()
circles(x = 0)

# Example 2:
plot0()
circles(x=-.5, radi=0.5, from=0.5*pi, to=0.25*pi)
circles(x=.5, radi=0.5, from=0.5*pi, to=0.25*pi, pie=TRUE)

# Example 3:
plot0()
circles(matrix(-.5+.5*stats::runif(18), ncol=3))

# Example 4:
plot0(x=c(-2,2),y=c(-2,2), asp=1)
circles(x=c(-1,1),c(1,1,-1,-1),from=pi*seq(0.25,1,by=0.25),to=1.25*pi, col=2, border=4, lwd=3)
}
\seealso{
\code{\link[graphics]{symbols}}, \code{\link[plotrix]{draw.circle}}, \code{\link[plotrix]{draw.arc}}.
}
\keyword{circle}
\keyword{geometries}
