% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorScale.R
\name{colorScale}
\alias{colorScale}
\title{Create a color scale}
\usage{
colorScale(
  x,
  y,
  col,
  at = NULL,
  labels = NULL,
  horiz = TRUE,
  percx = NULL,
  percy = NULL,
  labels.cex = 1,
  title = "legend",
  title.cex = 1.2
)
}
\arguments{
\item{x}{x-coordinates of the bottom-left corner of the legend.}

\item{y}{y-coordinates of the bottom-left corner of the legend.}

\item{col}{vector of colors.}

\item{at}{vector of integers indicating colors to be labeled.}

\item{labels}{labels.}

\item{horiz}{a logical. Should the color scale be horizontal?}

\item{percx}{size of the color scale along x axis.}

\item{percy}{size of the color scale along y axis.}

\item{labels.cex}{magnification to be used for labels.}

\item{title}{legend title.}

\item{title.cex}{magnification to be used for the title.}
}
\description{
Simple and flexible color scale.
}
\examples{
plot(c(0,10), c(0,10))
colorScale(5, 5, gpuPalette("cisl", 10), at = c(2, 4))
}
