% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compassRose.R
\name{compassRose}
\alias{compassRose}
\alias{compassRoseCardinal}
\title{Compass Rose}
\usage{
compassRose(x = 0, y = 0, rot = 0, cex.cr = 1, cex.let = 1,
  offset = 1.2, col = c(1, 8), border = c(1, 8), lty = 1)

compassRoseCardinal(x, y = x, rot = 0, cex.cr = 1, cex.let = 1,
  labels = c("S", "W", "N", "E"), offset = 1.2, col = c(1, 8),
  border = c(1, 8), lty = 1)
}
\arguments{
\item{x}{the x coordinates of the center of the compass rose.}

\item{y}{the y coordinates of the center of the compass rose.}

\item{rot}{rotation for the compass rose in degrees (clockwise).}

\item{cex.cr}{the magnification to be used for the whole compass rose.}

\item{cex.let}{the magnification to be used for labels.}

\item{offset}{label offset of the cardinal points.}

\item{col}{colors of the compass rose (see details).}

\item{border}{colors of the borders of the compass rose.}

\item{lty}{line type for the border lines.}

\item{labels}{labels for the cardinal direction.}
}
\description{
Draw a compass rose fully customizable.
}
\details{
This function draw a basic but fully personnalisable compassRose.
Note that There already exists one compass rose by Jim Lemon in \code{sp} package.
}
\section{Functions}{
\itemize{
\item \code{compassRose}: A compass rose with the four cardinal directions and additionnal directions.

\item \code{compassRoseCardinal}: A compass with the four cardinal directions only.
}}

\examples{
plot0()
compassRose(0, rot=25)
}
