% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compassRose1.R
\name{compassRose1}
\alias{compassRose1}
\alias{compassRose2}
\title{Compass Rose}
\usage{
compassRose1(x, y = x, rot = 0, cex.cr = 1, cex.let = 1,
  labels = c("S", "W", "N", "E"), offset = 1.2, col = c(1, 8),
  border = c(1, 8), lty = 1)

compassRose2(x = 0, y = 0, rot = 0, cex.cr = 1, cex.let = 1,
  offset = 1.2, col = c(1, 8), border = c(1, 8), lty = 1)
}
\arguments{
\item{x}{The x coordinates of the center of the compass rose.}

\item{y}{The y coordinates of the center of the compass rose.}

\item{rot}{Rotation for the compass rose in degrees (clockwise).}

\item{cex.cr}{The magnification to be used for the whole compass rose.}

\item{cex.let}{The magnification to be used for labels.}

\item{labels}{Labels for the cardinal direction.}

\item{offset}{Label offset of the cardinal points.}

\item{col}{Colors of the compass rose (see details).}

\item{border}{Colors of the borders of the compass rose.}

\item{lty}{Line type for the border lines.}

\item{percentage}{The percentage of the axis for which the values is returned.}
}
\description{
Draw a compass rose fully customizable.
}
\details{
This function draw a basic but fully personnalisable compassRose.
There already exists one compass rose by Jim Lemon in \code{sp} package.
}
\section{Functions}{
\itemize{
\item \code{compassRose1}: A compass rose with the four cardinal directions.

\item \code{compassRose2}: A compass rose with the four cardinal directions and additionnal directions.
}}
\examples{
plot0()
compassRose2(0, rot=25)
}

