% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastColors.R
\name{contrastColors}
\alias{contrastColors}
\alias{col2Hex}
\title{Contrast colors.}
\usage{
contrastColors(colors, how = "how_borw", alpha = FALSE)

col2Hex(colors, alpha = FALSE)
}
\arguments{
\item{colors}{vector of any of the three kinds of R color specifications, see \code{\link[grDevices]{col2rgb}}.}

\item{how}{a method to contrast \code{colors}. Methods currently available are \code{how_borw}, \code{how_cent}, \code{how_oppo} and \code{how_prop}, see details.}

\item{alpha}{logical value indicating whether the alpha channel (opacity) values should be returned.}
}
\description{
For a given set of colors, \code{contrastColors} returns an associated set
of colors.
}
\details{
Based on the sum of colors' saturation \code{how_borw} returns black or white,
\code{how_prop} proportionally remove or add some saturation. \code{how_oppo}
opposes the color (255-x) and \code{how_cent} centers the columns, i.e. remove
or add 127.
}
\section{Functions}{
\itemize{
\item \code{contrastColors}: Retuns a set of colors contrasted.

\item \code{col2Hex}: Returns the hexadecimal string associates to a given vector of colors.
}}

\examples{
#Example 1:
contrastColors('blue')
contrastColors('blue', how = 'how_prop')
}
\seealso{
\link[grDevices]{col2rgb}
}
\keyword{colors,}
\keyword{contrast}
