% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dfGantt}
\alias{dfGantt}
\title{dfGantt}
\format{A simulated data frame to create a Gantt chart.}
\usage{
data(dfGantt)
}
\description{
A data frame to showcase \code{\link[graphicsutils]{ganttChart}}.
}
\author{
David Beauchesne
}
\keyword{Gantt}
\keyword{datasets}
