% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse.R
\name{ellipse}
\alias{ellipse}
\title{Ellipse}
\usage{
ellipse(x = 0, y = x, mjradi = 1, mnradi = 0.5, from = 0, to = 2 *
  pi, rot = 0, incr = 0.01, pie = FALSE, ...)
}
\arguments{
\item{x}{the x coordinates of the centers of the ellipses.}

\item{y}{the y coordinates of the centers of the ellipses.}

\item{mjradi}{the major radii of the ellipses.}

\item{mnradi}{the minor radii of the ellipses.}

\item{from}{the angles, expressed in radians, from which ellipses are drawn.}

\item{to}{the angles, expressed in radians, to which ellipses are drawn.}

\item{rot}{the rotation angles (in degree) of the ellipses.}

\item{incr}{increments between two points to be linked (expressed in radians).}

\item{pie}{logical. If TRUE end points are linked with the center of the ellipse (default is set to FALSE).}

\item{...}{additional arguments to be passed to \code{\link{polygon}} function.}
}
\description{
Draw ellipses in a flexible way.
}
\details{
For a rotation angle of 0, major radii refer to the one along the x axis.

The number of cicrles drawn is given by the maximum argument length amng \code{x}, \code{y}, \code{radi}, \code{from} and \code{to} arguments.
Sizes are adjusted using \code{\link{rep_len}} function.

To plot ellipses, \code{\link{polygon}} function is called.
}
\note{
There is a similar function, called \code{draw.ellipse}, in the package \code{plotrix}.
}
\examples{
#Example 1:
plot0(asp=1)
ellipse()

#Example 2:
plot0()
for (i in seq(0,360,30)) ellipse(rot=i)

#Example 3:
plot0()
ellipse(matrix(-.5+.5*stats::runif(18), ncol=3))

#Example 4:
plot0(x=c(-2,2),y=c(-2,2), asp=1)
ellipse(x=c(-1,1),c(1,1,-1,-1),from=pi*seq(0.25,1,by=0.25),to=1.25*pi, col=2, border=4, lwd=3)
}
\keyword{ellipse}

