% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encircle.R
\name{encircle}
\alias{encircle}
\title{Encircle points}
\usage{
encircle(x, y = NULL, nb.pt = 20, off.set = 1, ...)
}
\arguments{
\item{x}{the x coordinates of a set of points. Alternatively, a
single argument \code{x} can be provided.}

\item{y}{the y coordinates of a set of points.}

\item{nb.pt}{the nuber of points to ge generated around eah coordinates.}

\item{off.set}{the y coordinates of a set of points.}

\item{...}{further arguments to be passed to \code{\link[graphics]{polygon}} function.}
}
\description{
Draw a polygons around a certains set of points.
}
\details{
The technique is fairly simple: for a set of coordinates x, y handled
using \code{xy.coords} a set of \code{nb.pt} number is generetaed at a
\code{off.set} distance of each coordinates, then a convex is drawn around
the coordinates using \link[grDevices]{chull}.
}
\examples{
coords <- cbind(runif(10), runif(10))
plot0(coords)
points(coords, bg='grey25', pch=21)
encircle(coords, border='#7b11a1', lwd=2)
}
