% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelop.R
\name{envelop}
\alias{envelop}
\title{Color the area between two lines.}
\usage{
envelop(x, upper, lower, ...)
}
\arguments{
\item{x}{vectors containing the x coordinates.}

\item{upper}{the y coordinates of the upper values.}

\item{lower}{the y coordinates of the lower values.}

\item{...}{additionnal arguments to be passed to \link[graphics]{polygon}.}
}
\description{
\code{envelop} colors the area between two series of y coordinates.
}
\details{
This function enables to color the areas between two set of values.
}
\examples{
plot0(c(0,10),c(0,10))
sz <- 100
seqx <- seq(0, 10, length.out=sz)
seqy1 <- 0.2*seqx*runif(sz, 0, 1)
seqy2 <- 4+0.25*seqx*runif(sz, 0, 1)
seqy3 <- 8+0.25*seqx*runif(sz, 0, 1)
envelop(seqx, seqy1, seqy2, col="grey85", border=NA)
envelop(seqx, seqy2, seqy3, col="grey25", border=NA)
}

