% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ganttChart.R
\name{ganttChart}
\alias{ganttChart}
\title{Gantt Chart}
\usage{
ganttChart(df, order = TRUE, add_milestone = order, axes = TRUE)
}
\arguments{
\item{df}{a data.frame (see details).}

\item{order}{a logical. Should the tasks be ordered? See below for more details.}

\item{add_milestone}{Should milestone be abbed? (not implemented yet)}

\item{axes}{a logical. Should the axes be added?}
}
\description{
A flexible Gantt Chart.
}
\details{
Data frame with the following column (order does not matter):
\itemize{
\item \code{milestone}
\item \code{due}
\item \code{start}
\item \code{task}
\item optionally a \code{status} and \code{col} columns
}
}
\references{
https://insileco.github.io/2017/09/20/gantt-charts-in-r/
}
\author{
David Beauchesne, Kevin Cazelles
}
