% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getangle2d.R
\name{getangle2d}
\alias{getangle2d}
\title{getangle2d}
\usage{
getangle2d(x0, y0, x1, y1, rad = FALSE)
}
\arguments{
\item{x0}{The x coordinates of points from which to}

\item{y0}{The y coordinates of points from which to}

\item{x1}{The x coordinates of points to which to draw.}

\item{y1}{The y coordinates of points to which to draw.}

\item{...}{Number of vertical lines drawn to select the size of the subplots regions.}
}
\description{
Calcul the angle of the vector between two points of a plane.
}
\examples{
#Example:
plot0(c(-10,10),c(-10,10))
arrows(c(0,3),c(0,0),c(3,0),c(3,-3))
cool <- getangle2d(c(0,3),c(0,0),c(3,0),c(3,-3))
}
\keyword{empty}
\keyword{plot}

