% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getIcon.R
\name{getIconNames}
\alias{getIcon}
\alias{getIconNames}
\title{getIcon}
\usage{
getIconNames()

getIcon(name, res = 256, destfile = NULL, col = NULL, quiet = FALSE,
  preview = FALSE, get_path = FALSE)
}
\arguments{
\item{name}{Name of the ocon to be dowloaded.}

\item{res}{The resolution of the icon to be downloaded.}

\item{destfile}{A character string with the name where the downloaded file is saved.  Tilde-expansion is performed.}

\item{col}{Color of the image.}

\item{quiet}{logical. If TRUE, no message are printed.}

\item{preview}{logical. If TRUE, the icon is displayed once downloaded.}

\item{get_path}{logical. If TRUE, the path to the file downloaded is returned. Otherwise a objet of class \code{nativeRaster} is returned.}
}
\description{
Draw a compass rose fully customizable.
}
\details{
The list of icons is available at \link{https://github.com/encharm/Font-Awesome-SVG-PNG/commit/161b9457c423f2dd53ce106360505ad0b286fa3e}
Resolutions available are 16, 22, 24, 32, 48, 64, 128 or 256.
To obtain the list of name that can be used in \code{getIcon}, use getIconNames().
If argument \code{destfile} is not specified, then the icon is stored in a temporary file.
Colour gradient as it may appear in original icon is not yet handled once the color is changed.
}
\section{Functions}{
\itemize{
\item \code{getIconNames}: Return the list of names available for downloading.

\item \code{getIcon}: Download an icon based on its name.
}}
\examples{
getIconNames()
getIcon(name="beer", col="grey80")
}

