% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gpu}
\alias{gpu}
\title{graphicsutils (gu) colors palettes}
\format{A list including six color palettes.}
\usage{
gpu
}
\description{
A collection of color palettes.
}
\author{
Kevin Cazelles
}
\keyword{datasets}
