% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuPalettes.R
\docType{data}
\name{gpuPalettes}
\alias{gpuPalettes}
\title{Complete list of graphisutils' color palettes}
\format{An object of class \code{list} of length 7.}
\usage{
gpuPalettes
}
\description{
Color palettes of the graphisutils package.
}
\keyword{datasets}
