% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradientPolygon.R
\name{gradientPolygons}
\alias{gradientPolygons}
\title{Gradient or an image.}
\usage{
gradientPolygons(x, y, color_palette = rainbow(100), image = NULL,
  res = 300, angle = 0, radian = FALSE, ...)
}
\arguments{
\item{x}{The x coordinates of points in the plot.}

\item{y}{The y coordinates of points in the plot.}

\item{image}{The y coordinates of points in the plot.}

\item{res}{The resolution of the graodient (ignored if image is not \code{NULL})}

\item{radian}{logical. If TRUE, then radian are used rather that degree.}

\item{...}{Additional arguments to be passed to \code{polygon}.}

\item{gradient}{The x coordinates of points in the plot.}
}
\description{
Draw a polygon colored by a color gradient or an image.
}
\details{
The function is based on the plot.defaut function. It simply makes the creation of an empty plot quicker
}
\examples{
# Example 1:
plot0(c(0,10),c(0,10))
}
\keyword{empty}
\keyword{plot}

