% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/homothety.R
\name{homothety}
\alias{homothety}
\title{Homothety}
\usage{
homothety(x, y, lambda, xcen = NULL, ycen = NULL, add = FALSE, ...)
}
\arguments{
\item{x}{The x coordinates of points. It can also be a matrix (see details).}

\item{y}{The y coordinates of points.}

\item{lambda}{The factor to be used for the homothetetic transformations.}

\item{xcen}{The x coordinate for the center of rotation.}

\item{ycen}{The y coordinate for the center of rotation.}

\item{add}{logical. If TRUE the set of transforned points are drawn as a polygon.}

\item{...}{Additionnal arguments to be passed to \code{polygon} function (used only if \code{add} is TRUE).}
}
\description{
Compute a homothetic transformation for a set of points. The transformed set of points is optionnaly drawn as a polygon.
}
\details{
Details about what is a homothetic transformation, see \url{https://en.wikipedia.org/wiki/Homothetic_transformation}.

If x is a matrix with more than 2 columns, then x is the first column and y the second one.

Note that \code{lambda}, \code{xcen} and \code{ycen} are unique meaning that \code{homothety} computes only one homothetic transformation.

Drawing the points computed is relevant only if there are more than 2 points.
}
\examples{
#Example:
plot0(c(0,10),c(0,10))
x <- c(4,6,5)
y <- c(2,2,4)
polygon(x,y)
poly2 <- homothety(x,y,2)
polygon(poly2$x,poly2$y)
poly3 <- homothety(x, y, -2.5, xcen=5, ycen=4, border=4, add=TRUE)
}
\keyword{homothethy}

