% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/howManyRC.R
\name{howManyRC}
\alias{howManyRC}
\title{How many rows and columns}
\usage{
howManyRC(n)
}
\arguments{
\item{n}{a positive integer (or coercible as is).}
}
\value{
A vector of two elements: the number of rows followed by the number of columns.
}
\description{
Compute the number of rows and columns to split a graphic window into panels
as equally as possible.
}
