% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactiveLayout.R
\name{interactiveLayout}
\alias{interactiveLayout}
\title{An interactive version of \code{layout}}
\usage{
interactiveLayout(n = 1, grain.x = 20, grain.y = grain.x, show = TRUE,
  now = TRUE)
}
\arguments{
\item{n}{Number of plot regions desired, default is 1.}

\item{grain.x}{Number of vertical lines drawn to select the size of the subplots regions.}

\item{grain.y}{Number of horizontal lines drawn to select the size of the subplots regions.}

\item{show}{logical. If TRUE \code{layout.show} is used to get a preview of the subplots regions.}

\item{now}{logical. If TRUE \code{layout} is called at \code{layout2} exit.}
}
\value{
the matrix use to draw the layout is returned as an invisible output.
}
\description{
This function provides an interactive version of the \code{layout} function. Once \code{layout2} is called, the users get a grid he can click on (thanks to \code{locator} function).
Once the clicks have been achieved, the users can get the matrix created and/or use directly layout and then call the function.
}
\details{
Arguments \code{grain.x} and \code{grain.y} control the aspect of the support
grid generated to locate the differents panels. Once the grid popup, the user
must click 2*n times to select the size of their n subplot. A panel is
delimited by two consecutive clicks: click 2*p and clicks 2*p+1. The two cells
the user clicked on will then be used to compute the area allocated for
panel p. Note that the user is allowed to use the same area for several plots
but only this area will actually be used only by the last one.

As \code{layout} is ultimately called, \code{layout2} has the same limits: currently 200 for the numbers of rows and columns and 10007 for the total number of cells.
}
\keyword{interactive}
\keyword{layout}
