% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.R
\name{panel}
\alias{panel}
\title{Add}
\usage{
panel(mat, dim = NULL, side = 1:4)
}
\arguments{
\item{mat}{a matrix object specifying the location of the next N figures on the output device.}

\item{dim}{dimension of the matrix to be created.}

\item{side}{the number of the side plot must be added.}
}
\description{
Return the values of either x-axis or y-axis for a given percentage.
}
\details{
This function intends to ease the positionning of additionnal marks such as text when axis have not common axis.
}

