% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pchimage.R
\name{pchimage}
\alias{pchimage}
\title{Images as ploting characters.}
\usage{
pchimage(x, y, obj = NULL, file = NULL, cex.x = 1, cex.y = cex.x,
  atcenter = TRUE, add = TRUE, ...)
}
\arguments{
\item{x}{The x coordinates of images to be drawn.}

\item{y}{The y coordinates of images to be drawn.}

\item{obj}{An object of class \code{nativeRaster} function.}

\item{file}{A path to either a \code{.png} file or a \code{.jpeg} file.}

\item{cex.x}{A numerical value giving the amount by which the horizontal width of the image should be magnified relative to the default.}

\item{cex.y}{A numerical value giving the amount by which the vertical width of the image should be magnified relative To the default.}

\item{atcenter}{logical. If TRUE x and y coordinates describe the center of the image. Otherwise they represent the bottom-left coordinates of the image.}

\item{add}{logical. Should images be added on the current graph ? If FALSE a new plot is created.}

\item{...}{Additional arguments to be passed to \code{rasterImage} function.}
}
\description{
\code{pchimage} returns a plot that displays an image. It enables to users to directly include a png or a jpeg file in a plot region by providing their path.
}
\details{
Either obj or file must be defined.

If a path is provided either \code{readPNG} or \code{readJPEG} according to the end of the file extension.
}
\examples{
#Example:
img <- readPNG(system.file("img", "Rlogo.png", package="png"), native=TRUE)
n <- 15
plot0(c(0,1),c(0,1))
pchimage(0.1+0.8*stats::runif(n), 0.1+0.8*stats::runif(n), cex.x=0.2+1.6*stats::runif(n), obj=img, angle=360*runif(n))
}
\keyword{image,}
\keyword{pch}
\keyword{plot,}

