% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percX.R
\name{percX}
\alias{percX}
\alias{percY}
\title{Values at a given percentage of the axis}
\usage{
percX(percentage = 90)

percY(percentage = 90)
}
\arguments{
\item{percentage}{The percentage of the axis for which the values is
returned.}
}
\description{
Returns the values of either x-axis or y-axis for a given percentage.
}
\details{
This function intends to ease the positioning of additional marks on plot
(such as annotation). Note that percentages refer to the entire plot window.
}
\section{Functions}{
\itemize{
\item \code{percX()}: Returns the values of x-axis for a given percentage.

\item \code{percY()}: Returns the values of y-axis for a given percentage.

}}
\examples{
plot0()
text(x = percX(90), y = percY(90), label = "cool")

}
