% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pickColors.R
\name{pickcolors}
\alias{pickcolors}
\title{Pick your colors up}
\usage{
pickcolors(colors = 6, ramp = rainbow(1024))
}
\arguments{
\item{colors}{The number of colors to be selected.}

\item{ramp}{A vector of colors.}
}
\description{
Generate an interactive interface to pick up colors and returns colors selected.
}
\details{
This colors selection open a grapical interface you navigate within following a red frame.
It allows you to confirm at each step wether or not you want to use the tone selected.
It does the same for the shades.
}
\keyword{color,}
\keyword{selection}

