% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot0.R
\name{plot0}
\alias{plot0}
\title{An empty plot function}
\usage{
plot0(x = c(-1, 1), y = x, fill = NULL, ...)
}
\arguments{
\item{x}{the x coordinates of points in the plot.}

\item{y}{the y coordinates of points in the plot.}

\item{fill}{The color to be used to fill the plot area.}

\item{...}{additional arguments to be passed to the defauls scatterplot
function \code{\link[graphics]{plot.default}}.}
}
\description{
\code{plot0} returns an a plot of a specific size without any symbols.
}
\examples{
# Example 1:
plot0()

# Example 2:
plot0(c(-10,10), asp=1)

# Example 2:
plot0(c(-10,10), asp=1, fill=8)
}
\seealso{
\code{\link[shape]{emptyplot}}
}
\keyword{empty}
\keyword{plot}
