% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMeans.R
\name{plotMeans}
\alias{plotMeans}
\title{plot a means an error asociated}
\usage{
plotMeans(formula, data, SE = FALSE, draw_axis = TRUE,
  col_err = par()$col, col_pt = par()$col, cex_pt = 1, ...)
}
\arguments{
\item{formula}{a formula, see \code{\link[stats]{formula}}.}

\item{data}{a data frame (or list) from which the variables in formula should be taken.}

\item{SE}{logical. If TRUE then the standard error is computed rather than the standard deviation.}

\item{draw_axis}{logical. If TRUE axes and box are drawn.}

\item{col_err}{color of the lines that reflect uncertainty.}

\item{col_pt}{color of the points that stand for means.}

\item{cex_pt}{magnification coefficient of the points that stand for means.}

\item{...}{Further graphical parameters (see \code{\link[graphics]{plot.default}} and ) may also be supplied as arguments, particularly, line type, \code{lty}, line width, \code{lwd}, color, \code{col} and for \code{type = "b"}, \code{pch}.  Also the line characteristics \code{lend}, \code{ljoin} and \code{lmitre}.}
}
\description{
Draw a Box around a Plot
}
\examples{
# Example 1:
dataset <- data.frame(dat=c(rnorm(50, 10, 2), rnorm(50, 20, 2)) , grp=rep(c("A","D"), each=50))
plotMeans(dat~grp, data=dataset, pch=15)
}
\keyword{deviation}
\keyword{means,}
\keyword{standard}

