% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyRange.R
\name{prettyRange}
\alias{prettyRange}
\title{Pretty range}
\usage{
prettyRange(x)
}
\arguments{
\item{x}{A vector of numerical values.}
}
\value{
A vector if two values that correspond to the range of values.
}
\description{
This function returns a pretty range of values for a given a vector of type \code{numeric}.
}
\details{
This function intends to generate range with round values.
}
\examples{
# Example 1:
vec <- stats::runif(20)
range(vec)
prettyRange(vec)
# Example 2:
prettyRange(c(3.85,3.88245))
}
\keyword{interval.}
\keyword{range,}
