% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhombi.R
\name{rhombi}
\alias{rhombi}
\title{Rhombi}
\usage{
rhombi(x, y = x, ldg = 1, sdg = ldg, rot = 0, area = FALSE, ...)
}
\arguments{
\item{x}{Vector of x coordinates of the centers of the losange}

\item{y}{Vector of y coordinates of the centers of the losange}

\item{ldg}{Vector of length of the large diagonals.}

\item{sdg}{Vector of length of the small diagonals.}

\item{rot}{The rotation angles (in degree) of the rhombi.}

\item{area}{logical. If TRUE the area of rhombis are returned.}

\item{...}{Additionnal arguments to be passed to polygon function.}
}
\description{
Add rhombi on a plot and optionnaly returns areas.
}
\details{
The number of rhombus maximual is provided by the length of the largest argument among x, y, ldg, sdg and rot.
Other arguments are repeated with the largest length as the desired one (see \code{rep_len}).
Additionnal arguments remain the same for every rhombus.
}
\examples{
# Example 1:
plot0(asp=1)
rhombi(0)

# Example 2:
plot0(c(-0.4,1.4),c(-0.4,1.4))
rhombi(runif(6), runif(5), runif(2), runif(3))

# Example 3:
plot0(asp=1)
rhombi(x=0, rot=seq(0,180,20), col=2, border=NA)
rhombi(x=0, rot=seq(0,180,30), ldg=0.6, col=7, border=NA)
}
\keyword{rhumbus}

