% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadowText.R
\name{shadowText}
\alias{shadowText}
\title{Shadow a Text}
\usage{
shadowText(x, y, labels, col = "white", bg = par()$fg, radius = 0.1, ...)
}
\arguments{
\item{x}{coordinate(s) of text(s) on the x-axis.}

\item{y}{coordinate(s) of text(s) on the y-axis.}

\item{labels}{A character vector to shadow.}

\item{col}{color of the text.}

\item{bg}{color of the background (shadow color).}

\item{radius}{Width of the shadow.}

\item{...}{others parameters to pass to \code{text()}.}
}
\description{
This function shadows a text and adds it to an existing plot.
}
\examples{
plot(1, type = "n", ann = FALSE, las = 1)
shadowText(x = 0.7, y = 1.3, labels = "This is a\nshadow text")
shadowText(x = 1.0, y = 1.3, labels = "This is a\nshadow text",
           family = "serif")
shadowText(x = 1.3, y = 1.3, labels = "This is a\nshadow text",
           family = "mono")
shadowText(x = 1.0, y = 1.0, labels = "This is a shadow text",
           family = "serif", cex = 3, col = "yellow", bg = "red")
shadowText(x = 0.7, y = 0.7, labels = "This is a\nshadow text",
           family = "serif", srt = 45)
shadowText(x = 1.0, y = 0.7, labels = "This is a\nshadow text",
           family = "serif", srt = 180)
shadowText(x = 1.3, y = 0.7, labels = "This is a\nshadow text",
           family = "serif", srt = -45)
}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@gmail.com}
}
\keyword{shadow}
\keyword{text,}
