% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showPalette.R
\name{showPalette}
\alias{showPalette}
\title{Show a color Palette}
\usage{
showPalette(x = grDevices::palette(), inline = FALSE, add_number = FALSE,
  add_codecolor = FALSE, cex_num = 1.2)
}
\arguments{
\item{x}{A vector of colors.}

\item{inline}{A logical. If TRUE, the colors are displayed on a single row.}

\item{add_number}{A logical. If TRUE, color vector's indices are added (default is set to FALSE).}

\item{add_codecolor}{A logical. If TRUE, the code color is displayed (default is set to FALSE).}

\item{cex_num}{The maginification coefficient of the color vector's indices.}
}
\description{
Generate an interactive interface to pick up colors and returns colors selected.
}
\details{
This function generates a graphical window splitted into 6 panels.
The top panel serves to select one tone. Currently channel alpha is ignored.
}
\examples{
showPalette(sample(1:100, 16), add_number=TRUE, add_codecolor = TRUE)
}
\keyword{color,}
\keyword{selection}
