% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stackedareas.R
\name{stackedareas}
\alias{stackedareas}
\title{Stacked areas chart.}
\usage{
stackedareas(val, index = NULL, rgy = 1, cumul = FALSE, transp = FALSE,
  legend = NULL, add = FALSE, col = NULL, pickcolors = FALSE, lty = 1,
  lwd = 1, border = NA, main = "", xlab = "", ylab = "")
}
\arguments{
\item{val}{A dataframe or a matrix containing a series of positive value, rows stand for popultaions.}

\item{index}{Values to be used for the x axis. Defaults set is}

\item{rgy}{A values that determines the range of y values. Default is set to 1 which means that the range of values is [0,1].}

\item{cumul}{logical. If TRUE, data are regarded as cumulative sums.}

\item{transp}{logical. If TRUE, the transpose of the data table is computed.}

\item{legend}{The text to be used as a legend for each area drawn.}

\item{add}{logical. Should the stacked arease be add on the current plot?}

\item{col}{Vector of colors, repeated if too small.}

\item{pickcolors}{logical. If TRUE, \code{\link{pickcolors}} is called to select colors.}

\item{lty}{The line type (see \code{\link{par}} documentation)}

\item{lwd}{The line width (see \code{\link{par}}q documentation)}

\item{border}{The color to draw the border. The default, ‘NULL’, means to 1use ‘par("fg")’.  Use ‘border = NA’ to omit borders.}

\item{main}{A main title for the plot, see also ‘title’.}

\item{xlab}{A label for the x axis, defaults to a description of ‘x’.}

\item{ylab}{A label for the y axis, defaults to a description of ‘y’}
}
\description{
Draw a stacked areas chart.
}
\details{
Areas are drawn using the \code{\link{polygon}} function and users can take advantage of ot to customize their stacked areas (using \code{lwd}, \code{lty} or \code{border} arguments).
}
\note{
The default colors have been inspired by four palettes found on line: \url{http:www.color-hex.com/color-palettes/}. \code{stackpoly} in \code{plotrix} package offers a valuable alternative.

Using a stacked areas chart with more than 20 areas should provide a figure really hard to read.
}
\examples{
## data for 8 populations at 25 different periods.
x <- data.frame(matrix(runif(200,2,10),8,25))
stackedareas(x)
# plot 1: default plot
stackedareas()

## plot 2:
stackedareas(x, index=2001:2025,rgy=100, border="white",pickcolors=TRUE, main="Stacked areas chart")

## plot 3: personalized plot
par(xaxs="i", yaxs="i", font=2, cex.axis=1.2, cex.lab=1.4, bty="l")
plot(c(1999,2027), c(-10,110), type="n", xlab="Years",ylab="Percentage", main="My customized stacked areas chart")
bgcol(col="#f2c4c4")
stackedareas(x, index=2001:2025, rgy=100, lwd=2, add=TRUE, border="transparent")
}
\keyword{circle,}
\keyword{plot}

