% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toFig.R
\name{toFig}
\alias{toFig}
\alias{toUser}
\title{Convert coordinates}
\usage{
toFig(x, y = NULL)

toUser(x, y = NULL)
}
\arguments{
\item{x}{the x coordinates of a set of points. Alternatively, a
single argument \code{x} can be provided.}

\item{y}{the y coordinates of a set of points.}
}
\description{
Convert user coordinates to figure region coordinates and conversely.
}
\details{
These functions can be used anytime, however, there are useful only once
\link[graphics]{plot.new} has been called.
}
\section{Functions}{
\itemize{
\item \code{toFig}: Convert figure region coordinates into user coordinates.

\item \code{toUser}: Convert user coordinates into figure region coordinates.
}}

