% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/translation.R
\name{translation}
\alias{translation}
\title{Translation.}
\usage{
translation(x, y, xtrans = 0, ytrans = 0, add = FALSE, ...)
}
\arguments{
\item{x}{The x coordinates of points. It can also be a matrix (see details).}

\item{y}{The y coordinates of points.}

\item{xtrans}{The x vector of the translation.}

\item{ytrans}{The y vector of the translation..}

\item{add}{logical. If TRUE the set of transforned points are drawn as a polygon.}

\item{...}{Additionnal arguments to be passed to \code{polygon} function (used only if \code{add} is TRUE).}
}
\description{
Compute a translation for a set of points. The transformed set of points is optionnaly drawn as a polygon.
}
\details{
Details about what is a translation, see \url{https://en.wikipedia.org/wiki/Translation_(geometry)}.

If x is a matrix with more than 2 columns, then x is the first column and y the second one.

Drawing the points computed is relevant only if there are more than 2 points.
}
\examples{
#Example 1:
plot0(c(0,10),c(0,10))
x <- c(4,6,5)
y <- c(2,2,4)
polygon(x,y)
trans1 <- translation(x,y,xtrans=2,ytrans=5, add=TRUE, border=4, lwd=2)

#Example 2:
x <- c(2,4,3,1)
y <- c(1,1,3,3)
plot0(c(0,10),c(0,10))
polygon(x,y)
for (i in 1:6) translation(x,y,xtrans=i,ytrans=i, add=TRUE, border=i+1, lwd=2)
}
\keyword{translation}

