% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vecfield2d.R
\name{vecfield2d}
\alias{vecfield2d}
\title{Vector Fields.}
\usage{
vecfield2d(seqx, seqy, FUN, args, cex.x = 0.25, cex.y = 0.25, log = TRUE,
  add = FALSE, ...)
}
\arguments{
\item{seqx}{The x coordinates of the set of vector.}

\item{seqy}{The y coordinates of the set of vector.}

\item{FUN}{The function that describes the dynamical system (see details).}

\item{args}{The parameters of the dynamical system (see details).}

\item{cex.x}{The magnification coefficient to be used for the x coordinates of the vectors.}

\item{cex.y}{The magnification coefficient to be used for the y coordinates of the vectors.}

\item{log}{logical. If TRUE, the lenghts of arrows are log-transformed.}

\item{add}{logical. If TRUE, the vector field is added on the current plot.}

\item{...}{Additionnal arguments to be passed to \code{arrows2}.}
}
\description{
Draw a vector field associated to a dynamical system of dimension 2.
}
\details{
The \code{FUN} function to be used must be a function of two arguments. The first argument must contain the dynamical variables
as a vector of two elements. Values in \code{seqx} and \code{seqy} will be used as the values of the dynamical variables.
The second arguments must contain all the other parameters that shapes the dynamical system.
}
\examples{
# Example:
systLin <- function(X, beta){
    Y <- matrix(0,ncol=2)
    Y[1] <- beta[1]*X[1]+beta[2]*X[2]
    Y[2] <- beta[3]*X[1]+beta[4]*X[2]
    return(Y)
}
seqx <- seq(-2,2,0.35)
seqy <- seq(-2,2,0.35)
# Plot 1:
vecfield2d(seqx, seqy, FUN=systLin, args=c(0,-1,1,0))
# Plot 2:
vecfield2d(seqx, seqy, FUN=systLin, args=c(0,-1,1,0), log=FALSE, cex.hh=1.4, cex.sk=0.8, col=8)
}
\keyword{empty}
\keyword{plot}

