% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAxis.R
\name{addAxis}
\alias{addAxis}
\title{Add Spatial Cardinal Directions to Axis Labels}
\usage{
addAxis(side, at = axTicks(side), mgp = par()$mgp, digits = 0, ...)
}
\arguments{
\item{side}{an integer between 1 (bottom x-Axis) and 4 (right y-Axis).}

\item{at}{values on the axis to add labels.}

\item{mgp}{customize axis positions. See \code{par()}.}

\item{digits}{number of digits of labels.}

\item{...}{other parameters to pass to \code{axis()}.}
}
\description{
This function adds cardinal directions to axis labels in a map.
}
\examples{
maps::map()
addFrame(grid = TRUE, add = TRUE, width = 5)
addAxis(1)
addAxis(2, at = seq(-50, 50, 100), digits = 2)
addAxis(3, lwd = 0, lwd.ticks = 0.5)
addAxis(4, las = 1, col.axis = "red")

}
\author{
Nicolas CASAJUS, \email{nicolas.casajus@gmail.com}
}
\keyword{axis,}
\keyword{spatial}
