% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFrame.R
\name{addFrame}
\alias{addFrame}
\title{Add a Checkerboard-Like Frame around a Plot}
\usage{
addFrame(
  at_x,
  at_y,
  col = c("white", border),
  border = "black",
  width = NULL,
  lwd = 1,
  lty = 1,
  asp = NA,
  grid = FALSE,
  add = FALSE,
  ...
)
}
\arguments{
\item{at_x}{positions on the x-Axis of the borders of rectangles.}

\item{at_y}{positions on the y-Axis of the borders of rectangles.}

\item{col}{one or several colors of rectangles background.}

\item{border}{color of rectangles border.}

\item{width}{width of the frame.}

\item{lwd}{thickness of rectangles border (and grid).}

\item{lty}{type of rectangles border lines (and grid).}

\item{asp}{aspect ratio of the plot (see \code{plot.window()}).}

\item{grid}{a boolean. If \code{TRUE} adds vertical and horizontal lines at \code{at_x}
and \code{at_y}.}

\item{add}{a boolean. If \code{TRUE} adds the frame to an existing plot.}

\item{...}{other graphical parameters as in \code{par()}.}
}
\description{
This function adds a checkerboard-like frame around an existing plot (or a
new one) as an ensemble of small rectangles whose colors, position and width
can be specified by user.
}
\details{
If \code{add = FALSE}, a new plot is drawn with \code{xlim = ylim = c(-1, 1)}. Argument
\code{col} can be set with one or multiple colors. If only one color is specified
in \code{col}, border color will be used to draw the checkerboard. If \code{width} is
\code{NULL}, the width of the frame is equivalent to 4\% of the extent of the plot
area.
}
\examples{
## Grid and graduations
addFrame()
addFrame(grid = TRUE)
addFrame(grid = TRUE, lwd = 0.25, lty = 3)
addFrame(at_x = seq(-1, 1, by = 0.1))

## Frame width
addFrame(width = 0.25)
addFrame(at_x = seq(-1, 1, by = 0.01), width = 1.00)

## Colors
addFrame(col = 1:6)
addFrame(col = "lightgray", border = "darkgray")

## Adding to a plot
maps::map()
addFrame(grid = TRUE, add = TRUE, width = 5)
maps::map.axes()

## Other graphical parameters
addFrame(bg = "darkgray", xaxs = "i", yaxs = "i")
}
\author{
Nicolas Casajus, \email{nicolas.casajus@gmail.com}
}
\keyword{box,}
\keyword{frame,}
\keyword{grid}
