% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/arrows2.R
\name{arrows2}
\alias{arrows2}
\title{Add arrows to a plot}
\usage{
arrows2(x0, y0, x1 = x0, y1 = y0, cex.hg = 1, cex.hh = 1,
  pct.hw = 0.1, cex.sk = 1, space = 1, twoheaded = FALSE, ...)
}
\arguments{
\item{x0}{The x coordinates of points from which to draw.}

\item{y0}{The y coordinates of points from which to draw.}

\item{x1}{The x coordinates of points to which to draw.}

\item{y1}{The y coordinates of points to which to draw.}

\item{cex.hg}{The magnification coefficient to be used for the height of the arrows.}

\item{cex.hh}{The magnification coefficient to be used for the height of the heads of the arrows.}

\item{pct.hw}{The length of the heads of the arrows expressed as a percentage of the total length.}

\item{cex.sk}{The magnification coefficient to be used to chance the height of the arrows towards the heads.}

\item{space}{The magnification coefficient to be used for the space between points between which arrows are drawn and the arrows them-selves.}

\item{twoheaded}{logical. If TRUE two headed arrows are drawn.}

\item{...}{Additional arguments to be passed to \code{polygon} function.}
}
\description{
Draw arrows between pairs of points.
}
\details{
Arrows as a polygons
}
\examples{
# Example 1:
plot0(c(0,10),c(0,10))
points(c(2,6),c(8,8), col=8, pch=19)
arrows2(2,2,x1=2,y1=4)
arrows2(2,c(4,6,8),x1=6,y1=c(4,6,8), col=8, space=1.2)

# Example 2:
plot0(c(0,10),c(0,10))
arrows2(1,8,x1=9,y1=6, lwd=2, col=4, border=2, lty=2)
arrows2(1,6,x1=9,y1=4, lwd=2, cex.hg=2, cex.hh=2, cex.sk=0.5, col=4, border=2, lwd=2)
arrows2(1,4,x1=9,y1=2, lwd=2, cex.hg=2, cex.hh=2, cex.sk=0.5, pct.hw=0.18, twoheaded=TRUE,
col=4, border=2, lwd=2)
}
\keyword{arrows}

