% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrows2.R
\name{arrows2}
\alias{arrows2}
\title{Add arrows to a plot.}
\usage{
arrows2(x0, y0, x1 = x0, y1 = y0, off0 = 0, off1 = off0,
  cex.arr = 1, cex.shr = 1, cex.hh = 1, cex.hl = 1,
  prophead = TRUE, twoheaded = FALSE, ...)
}
\arguments{
\item{x0}{the x coordinates of points from which to draw arrows.}

\item{y0}{the y coordinates of points from which to draw arrows.}

\item{x1}{the x coordinates of points to which to draw arrows.}

\item{y1}{the y coordinates of points to which to draw arrows.}

\item{off0}{offset of points from which to draw arrows.}

\item{off1}{offset of points to which to draw arrows.}

\item{cex.arr}{the magnification coefficient to be used for the heights of the arrows.}

\item{cex.shr}{the magnification coefficient to be used to change the height of arrows towards their heads.}

\item{cex.hh}{the magnification coefficient to be used for the heights of arrows' head.}

\item{cex.hl}{the magnification coefficient to be used for the lengths of arrows' head.}

\item{prophead}{logical. If TRUE arrows are drawn with head proportional to the length of the arrows.}

\item{twoheaded}{logical. If TRUE two-headed arrows are drawn, default is FALSE.}

\item{...}{additional arguments to be passed to \code{polygon} function.}
}
\description{
Draw a custom arrows between pairs of points.
}
\examples{
# Example 1:
plot0(c(0,10),c(0,10))
arrows2(1,9,8)
arrows2(1,8,8,1,cex.hh=1.2, cex.hl=1.2, col='grey30', lwd=1.2, prophead=TRUE, twoheaded=TRUE)
arrows2(5,9,5,1)

# Example 2:
plot0(c(0,1),c(0,1))
arrows2(runif(2), runif(2), x1=runif(2), y1=runif(2))
arrows2(runif(2), runif(2), x1=runif(2), y1=runif(2), prophead=FALSE, lty=3)
}
\seealso{
\code{[graphics::arrows()]}, \code{[shape::Arrows()]}
}
\keyword{arrows}
