% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bgcol.R
\name{bgcol}
\alias{bgcol}
\title{Background color}
\usage{
bgcol(...)
}
\arguments{
\item{...}{Additional arguments to be passed to \code{\link{rect}} function.}
}
\description{
Color the background of the plot region.
}
\details{
The function calls \code{\link{rect}} and draw a colored (defaut is light blue) rectangle whose dimension are given par argument of function \code{\link{par}}.
}
\note{
In par(), the argument bg colors all the window.
}
\examples{
#Example 1:
plot0()
bgcol()

#Example 2:
plot0()
bgcol(col=8, lwd=4, border=4)
}
\keyword{background,}
\keyword{color}

