% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biBoxplot.R
\name{biBoxplot}
\alias{biBoxplot}
\title{An boxplot of two distributions.}
\usage{
biBoxplot(dist1, dist2)
}
\arguments{
\item{boxplot}{}
}
\description{
Double boxplot
}
\details{
Do not assess the distributions. Based on quantile only.
}
\examples{
# Example 1:
}
\seealso{
\code{\link[graphics]{box}}
}
\keyword{box}
