% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biBoxplot.R
\name{biBoxplot}
\alias{biBoxplot}
\title{A bi-boxplot}
\usage{
biBoxplot(
  df1,
  df2 = df1,
  probs = c(0.01, 0.25, 0.5, 0.75, 0.99),
  width = 0.2,
  sta_wd = 0.5,
  median = NULL,
  staples = NULL,
  whiskers = NULL,
  col_left = "grey75",
  col_right = col_left,
  add = FALSE,
  at = NULL
)
}
\arguments{
\item{df1, df2}{first and second set of boxplots.}

\item{probs}{numeric vector of five probabilities, see \code{\link[stats:quantile]{stats::quantile()}}.}

\item{width}{a vector giving the relative widths of the boxes making up the plot.}

\item{sta_wd}{staple width.}

\item{median}{a list of arguments passed to \code{\link[graphics:lines]{graphics::lines()}} to custom the median line.}

\item{staples}{a list of arguments passed to \code{\link[graphics:lines]{graphics::lines()}} to custom the staples.}

\item{whiskers}{a list of arguments passed to \code{\link[graphics:lines]{graphics::lines()}} to custom the whiskers.}

\item{col_left}{color of the left boxes.}

\item{col_right}{color of the right boxes.}

\item{add}{a logical. Should the biboxplots be added on the current graph? If \code{FALSE} then a new plot is created.}

\item{at}{numeric vector giving the locations where the boxplots should be drawn. Same default behavior as in \code{\link[graphics:boxplot]{graphics::boxplot()}}.}
}
\description{
Draws boxplots and bi-boxplots.
}
\details{
This functions does not attempt to assess the distributions as it is based on quantiles.
}
\examples{
dis1 <- list(stats::rnorm(1000, sd = 0.25))
biBoxplot(dis1)
}
\keyword{boxplots}
