% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box2.R
\name{box2}
\alias{box2}
\title{Box version 2}
\usage{
box2(side = 1:4, which = "plot", col2fill = NULL, ...)
}
\arguments{
\item{side}{A numerical/character vector or a character string specifying which side(s) of the plot/figure the box is to be drawn (see details).}

\item{which}{A character, one of \code{plot}, \code{figure}, \code{inner} and \code{outer}.}

\item{col2fill}{The color to be used to fill the box.}

\item{...}{Further graphical parameters (see \code{\link[graphics]{par}}) may also be supplied as arguments, particularly, line type, \code{lty}, line width, \code{lwd}, color, \code{col} and for \code{type = "b"}, \code{pch}.  Also the line characteristics \code{lend}, \code{ljoin} and \code{lmitre}.}
}
\description{
Draw a Box around a Plot
}
\details{
This function intends to give more flexibility to the \code{\link[graphics]{box}} function.
As \code{which} parameter, users provides an object first coerced by ‘as.character’ to a character string that is secondly split into single characters. For all of these characters, matches are sought with all elements of \code{1, 2, 3, 4, b, l, t, r} where 1=below, 2=left, 3=above, 4=right, b=below, l=left, t=above and r=right.
}
\examples{
# Example 1:
plot0()
box2()
##
plot0()
box2("14", col2fill="grey80", lwd=2)
plot0()
box2(c(1,4), col2fill="grey80", lwd=2)
plot0()
box2(c(1,4), col2fill="grey80", lwd=2)
# Example 2:
par(mfrow=c(2,2),oma=c(2,2,2,2))
plot(0,0)
box("outer",lwd=2)
box("inner",lwd=2)
plot(0,0)
plot(0,0)
plot(c(0,1),c(0,2))
box2(which="figure",lwd=2, col2fill=2)
box2(side=12, lwd=2, col2fill=8)
points(c(0,1),c(0,2))
axis(2)
axis(1)
title(xlab="coolx", ylab="cooly")
}
\seealso{
\code{\link[graphics]{box}}
}
\keyword{box}

