% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box2.R
\name{box2}
\alias{box2}
\title{Alternative box function}
\usage{
box2(side, which = c("plot", "figure", "outer", "inner"), fill = NULL, ...)
}
\arguments{
\item{side}{a numerical or character vector or a character string
specifying which side(s) of the plot the box is to be drawn (see details).}

\item{which}{a character, one of \code{plot}, \code{figure}, \code{inner} and \code{outer}.}

\item{fill}{the color to be used to fill the box.}

\item{...}{further graphical parameters (see \code{\link[graphics:par]{graphics::par()}}) may also be
supplied as arguments, particularly, line type, \code{lty}, line width, \code{lwd},
color, \code{col} and for \code{type = 'b'}, \code{pch}. Also the line characteristics
\code{lend}, \code{ljoin} and \code{lmitre}.}
}
\description{
Draw a box around a plot.
}
\details{
This function intends to give more flexibility to the
\code{\link[graphics:box]{graphics::box()}} function. As \code{which} parameter, the user provides an
object first coerced by \code{as.character} to a character string that is
secondly split into single characters. For all of these characters, matches
are sought with all elements of \verb{1, 2, 3, 4, b, l, t, r} where \verb{1=below, 2=left, 3=above, 4=right, b=below, l=left, t=above and r=right}.
}
\examples{
# Example 1:
plot0()
box2()

# Example 2:
plot0()
box2("14", fill = "grey80", lwd = 2)
plot0()
box2(c(1, 4), fill = "grey80", lwd = 2)
plot0()
box2(c(1, 4), fill = "grey80", lwd = 2)

# Example 3:
par(mfrow = c(2, 2), oma = c(2, 2, 2, 2))
plot0(0, 0)
box("outer", lwd = 2)
box("inner", lwd = 2)
plot.default(0, 0)
plot.default(0, 0)
plot0()
box2(which = "figure", lwd = 2, fill = 2)
box2(side = 12, lwd = 2, fill = 8)

}
\seealso{
\code{\link[graphics:box]{graphics::box()}}
}
\keyword{box}
