% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/circle.R
\name{circle}
\alias{circle}
\title{Draw circles}
\usage{
circle(x = 0, y = x, radi = 1, from = 0, to = 2 * pi, incr = 0.01,
  pie = FALSE, ...)
}
\arguments{
\item{x}{The x coordinates of the centers of the circles.}

\item{y}{The y coordinates of the centers of the circles.}

\item{radi}{The radii of the circles.}

\item{from}{The angles, expressed in radians, from which circles are drawn.}

\item{to}{The angles, expressed in radians, to which circles are drawn.}

\item{incr}{Increments between two points to be linked (expressed in radians).}

\item{pie}{logical. If TRUE end points are linked with the center of the circle (default is set to FALSE).}

\item{...}{Additional arguments to be passed to \code{\link{polygon}} function.}
}
\description{
Draw circles in a flexible way.
}
\details{
The number of circles drawn is given by the maximum argument length amng \code{x}, \code{y}, \code{radi}, \code{from} and \code{to} arguments.
Sizes are adjusted using \code{\link{rep_len}} function.

To plot circles, \code{\link{polygon}} function is called.
}
\examples{
#Example 1:
plot0()
circle()

#Example 2:
plot0()
circle(x=-.5, radi=0.5, from=0.5*pi, to=0.25*pi)
circle(x=.5, radi=0.5, from=0.5*pi, to=0.25*pi, pie=TRUE)

#Example 3:
plot0()
circle(matrix(-.5+.5*stats::runif(18), ncol=3))

#Example 4:
plot0(x=c(-2,2),y=c(-2,2), asp=1)
circle(x=c(-1,1),c(1,1,-1,-1),from=pi*seq(0.25,1,by=0.25),to=1.25*pi, col=2, border=4, lwd=3)
}
\keyword{circle}

