% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compassRose.R
\name{compassRose}
\alias{compassRose}
\alias{compassRoseCardinal}
\title{Compass Rose}
\usage{
compassRose(x = 0, y = 0, labels = c("S", "W", "N", "E"), rot = 0,
  cex.cr = 1, cex.let = cex.cr, col.cr = c(1, 8), col.let = 1,
  border = c(1, 8), offset = 1.2, add = TRUE, ...)

compassRoseCardinal(x, y = x, rot = 0, cex.cr = 1, cex.let = 1,
  labels = c("S", "W", "N", "E"), offset = 1.2, col.cr = c(1, 8),
  col.let = 1, border = c(1, 8), ...)
}
\arguments{
\item{x}{the x coordinates of the center of the compass rose.}

\item{y}{the y coordinates of the center of the compass rose.}

\item{labels}{a vector of four character strings used as labels for the cardinal directions.}

\item{rot}{rotation for the compass rose in degrees (clockwise).}

\item{cex.cr}{the magnification to be used for the whole compass rose.}

\item{cex.let}{the magnification to be used for labels.}

\item{col.cr}{a vector of colors used to draw compass rose (see details).}

\item{col.let}{a character string specifying the labels' color.}

\item{border}{a vector of colors of the borders of the compass rose.}

\item{offset}{label offset of the cardinal points.}

\item{add}{a logical. Should the compass rose be added on the current graph?}

\item{...}{additional arguments to be passed to \code{[graphics::polygon()]}.}
}
\description{
This function draws a fully personnalisable compass rose.
}
\details{
Both \code{col.cr} and \code{border} are repeated over (\code{[base::rep()]}
is called) so it has a 8 elements, meaning all triangles the compass rose is
made of could have their own color.

Note that there already exists a similar function by Jim Lemon in \code{sp} package.
}
\section{Functions}{
\itemize{
\item \code{compassRose}: A compass rose with the four cardinal directions and additionnal directions.

\item \code{compassRoseCardinal}: A compass with the four cardinal directions only.
}}

\examples{
compassRose(0, rot=25, cex.cr = 2, col.let =2, add = FALSE)
}
