% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darken.R
\name{ramp}
\alias{darken}
\alias{lighten}
\alias{ramp}
\title{Lightened or darkened colors}
\usage{
ramp(fromcol, tocol, percentage = 50, as_rgb = FALSE)

darken(col, percentage = 50, as_rgb = FALSE)

lighten(col, percentage = 50, rgb = FALSE)
}
\arguments{
\item{fromcol}{The starting color.  color to be changed.}

\item{tocol}{The color towards which the 'fromcolor' will be nuanced.}

\item{percentage}{percentage determining to what extent the color .}

\item{as_rgb}{A logical indicating wheteher color must be returned as a matrix object.}

\item{col}{the color to be darkened or lightened.}
}
\description{
Return lightened or darkened colors, vetorized over \code{percentage}.
\code{ramp} is the general function any couple of colors. \code{darken} and
\code{lightened} calls \code{ramp} to darken and lighten a given color.
}
\section{Functions}{
\itemize{
\item \code{ramp}: A color standing between two given hues.

\item \code{darken}: A darkened color.

\item \code{lighten}: A lightened color.
}}
\examples{
darken('red', 50)
somereds <- lighten('red', seq(10,90,9))
showPalette(somereds)
}
\seealso{
\code{\link{colorRampPalette}}
}

