% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darken.R
\name{ramp}
\alias{ramp}
\alias{darken}
\alias{lighten}
\title{Lighten or darken colors}
\usage{
ramp(fromcol, tocol, percentage = 50, as_rgb = FALSE)

darken(col, percentage = 50, as_rgb = FALSE)

lighten(col, percentage = 50, as_rgb = FALSE)
}
\arguments{
\item{fromcol}{starting color, i.e. if \code{percentage = 0}, it is the color returned.}

\item{tocol}{color to nuance \code{fromcol}, i.e. if \code{percentage = 100}, it is the color returned.}

\item{percentage}{percentage determining the percentage of \code{tocol} used to nuance \code{fromcol}.}

\item{as_rgb}{a logical. Should the color(s) returned as a matrix object?}

\item{col}{the color to be darkened or lightened.}
}
\description{
Returns lightened or darkened colors, vectorised over \code{percentage}.
\code{ramp} is valid for any couple of colors. Functions \code{darken} and
\code{lightened} actually call \code{ramp} to darken and lighten a given color.
}
\section{Functions}{
\itemize{
\item \code{ramp}: Returns a shaded color.

\item \code{darken}: Returns a darkened color.

\item \code{lighten}: Returns a lightened color.
}}

\examples{
showPalette(ramp("blue", "red", 10*3:7))
darken('red', 50)
more_reds <- lighten('red', seq(10,90,9))
showPalette(more_reds)
}
\seealso{
\code{\link{colorRampPalette}}
}
