% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frameit.R
\name{frameit}
\alias{frameit}
\title{Draw a frame}
\usage{
frameit(nbc = 10, cex.x = 1, cex.y = cex.x, col = c("grey45", "grey85"),
  border = NA)
}
\arguments{
\item{nbc}{number of rectangles drawn for each axis.}

\item{cex.x}{control the dimension}

\item{cex.y}{The radii of the circles.}

\item{col}{color of the rectangles drawn.}

\item{border}{color of the borders if the rectangles drawn.}
}
\description{
Draw a frame around the plot region. \code{frameit} function draw an ensemble of small rectangles around the plot region to create a frame which colors can be specified.
}
\details{
The number of rectangles could be different from \code{nbc} as \code{pretty} is called to position them.

The 'col' argument determines the succession of colors to be applied to each axis.
}
\examples{
#Example 1:
plot0()
frameit()

#Example 2:
plot0()
frameit(cex.x=1.5, col=c(2,3), border=1)
}
\keyword{frame}

