% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ganttChart.R
\name{ganttChart}
\alias{ganttChart}
\title{Gantt Chart}
\usage{
ganttChart(df, col = gpuPalette("insileco")[-1], perc_txt = 20,
  order = TRUE)
}
\arguments{
\item{df}{a data.frame (see details),}

\item{col}{color palette.}

\item{perc_txt}{percentage of the wodth dedicated to the description of tasks.}

\item{order}{a logical should the tasks be ordered.}
}
\description{
Gant Chart.
}
\examples{
data(dfGantt)
ganttChart(dfGantt)
}
\author{
David Beauchesne
}
