% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ganttChart.R
\name{ganttChart}
\alias{ganttChart}
\title{Gantt Chart}
\usage{
ganttChart(
  df,
  task_order = TRUE,
  mstone_add = task_order,
  mstone_spacing = 1,
  mstone_lwd = 2,
  axes = TRUE,
  mstone_font = 2,
  lighten_done = 0
)
}
\arguments{
\item{df}{a data.frame (see details).}

\item{task_order}{a logical. Should the tasks be ordered? See below for more details.}

\item{mstone_add}{Should milestones be added?}

\item{mstone_spacing}{spacing between milestones (expressed as figure unit).}

\item{mstone_lwd}{lines width for the milestone.}

\item{axes}{a logical. Should the axes be added?}

\item{mstone_font}{font of milestone (ignore of \code{axes} is \code{FALSE}).}

\item{lighten_done}{percentage use to lighten done task (see \code{\link[=lighten]{lighten()}}).
Default set to 0, so the completed task of a given milestone have the same
color as the pending ones.}
}
\description{
A flexible Gantt Chart.
}
\details{
Argument \code{df} should be a data frame with the following columns (in any order):
\itemize{
\item \code{milestone}: milestones names,
\item \code{due}: due date (will be converted into a date with \code{\link[=as.Date]{as.Date()}}),
\item \code{start}: start date (will be converted into a date with \code{\link[=as.Date]{as.Date()}}),
\item \code{task}: tasks names.
It might as well include any of the following optional columns:
\item \code{done}: vector of logicals indicating whether the task if completed
\item \code{col}: to custom the color of the tasks.
}
}
\examples{
ff <- ganttChart(dfGantt, mstone_lwd = 3, mstone_spacing = 0.6,
  lighten_done = 80)
}
\references{
\url{https://insileco.github.io/2017/09/20/gantt-charts-in-r/}
}
\author{
David Beauchesne, Kevin Cazelles
}
