% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getangle2d.R
\name{getAngle2d}
\alias{getAngle2d}
\title{Angles between vectors.}
\usage{
getAngle2d(x0, y0, x1, y1, rad = FALSE)
}
\arguments{
\item{x0, y0}{the x and y coordinates of the first set of vector.}

\item{x1, y1}{the x and y coordinates of the second set of vector.}

\item{rad}{a logical. Should angles be expressed radians?}
}
\description{
Calcul the angles of two set of vectors.
}
\examples{
# Example:
plot0(c(-10, 10), c(-10, 10))
arrows(c(0, 3), c(0, 0), c(3, 0), c(3, -3))
cool <- getAngle2d(c(0, 3), c(0, 0), c(3, 0), c(3, -3))

}
\keyword{angles,}
\keyword{vectors}
