% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuPalettes.R
\name{gpuPalette}
\alias{gpuPalette}
\title{graphicsutils colors palettes}
\usage{
gpuPalette(id, ncol)
}
\arguments{
\item{id}{either an integer or a character string corresponding names of desired color palettes; if missing then color palettes available are prompted.}

\item{ncol}{An integer giving the number of colors to be returned (involves a call to \code{\link[grDevices:colorRampPalette]{grDevices::colorRampPalette()}}). If missing, then the entire palette is returned.}
}
\value{
A vector of character strings of the hexadecimal colors values.
}
\description{
A couple of color palettes including our inSileco palette and a couple
inspired by popular websites.
}
\examples{
showPalette(gpuPalette(c('atom', 'cisl')))
showPalette(gpuPalette('cisl', 100))
}
\keyword{colors,}
\keyword{palette}
