% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuPalettes.R
\name{gpuPalette}
\alias{gpuPalette}
\title{graphicsutils colors palette}
\usage{
gpuPalette(names = NULL, ncol = NULL)
}
\arguments{
\item{names}{names of desired color palettes; if \code{NULL}, color palettes available are printed.}

\item{ncol}{An integer indicating the number of colors to be returned (involves a call to \code{\link[grDevices:colorRampPalette]{grDevices::colorRampPalette()}}). Default is set to \code{NULL} in which case the number in the one the palette is made of.}
}
\value{
A vector of character strings of the hexadecimal colors values.
}
\description{
A couple of color palettes mainly insipered by popular websites.
}
\examples{
showPalette(gpuPalette(c('atom', 'cisl')))
showPalette(gpuPalette('cisl', 100))
}
\keyword{colors,}
\keyword{palette}
