% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuPalettes.R
\name{gpuPalette}
\alias{gpuPalette}
\title{graphicsutils colors palette}
\usage{
gpuPalette(names = NULL)
}
\arguments{
\item{names}{Names of desired palette; if \code{NULL}, color palettes available are printed.}
}
\value{
A vector of colours.
}
\description{
A couple of color palettes mainly insipered by popular websites.
}
\examples{
showPalette(gpuPalette(c('atom', 'cisl')))
}
\keyword{color}
\keyword{palette}
