% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactiveLayout.R
\name{interactiveLayout}
\alias{interactiveLayout}
\title{An interactive version of \code{layout}}
\usage{
interactiveLayout(
  n = 1,
  grain.x = 20,
  grain.y = grain.x,
  show = TRUE,
  now = TRUE
)
}
\arguments{
\item{n}{Number of plot regions, default is set to 1.}

\item{grain.x}{Number of vertical lines drawn to select the size of the
subplots regions.}

\item{grain.y}{Number of horizontal lines drawn to select the size of the
subplots regions.}

\item{show}{logical. If \code{TRUE} \code{\link[graphics:layout]{graphics::layout.show()}} is used to get a
preview of the subplots regions.}

\item{now}{logical. If \code{TRUE} \code{\link[graphics:layout]{graphics::layout()}} is called on exit.}
}
\value{
the matrix use to draw the layout is returned as an invisible output.
}
\description{
\code{layout2()} provides an interactive version of \code{\link[graphics:layout]{graphics::layout()}}. Once
called, an interactive grid is displayed and can be used to create a layout.
}
\details{
Arguments \code{grain.x} and \code{grain.y} control the aspect of the support
grid generated to locate the different panels. Once the grid pops up, the
user must click 2\emph{n times to select the size of the n subplots. A panel is
delimited by two consecutive clicks, the tow cells selected by click 2}p
and clicks 2*p+1  will be used to compute the area allocated to panel p.
Note that the user is allowed to use the same area for several plots
but only this area will actually be used only by the last one.

As \code{\link[graphics:layout]{graphics::layout()}} is ultimately called, \code{layout2()} has the same
limits, i.e. currently 200 for the numbers of rows and columns and 10007
for the total number of cells.
}
\keyword{interactive}
\keyword{layout}
