% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pchimage.R
\name{pchImage}
\alias{pchImage}
\title{Images as ploting characters.}
\usage{
pchImage(x, y, obj = NULL, file = NULL, cex.x = 1, cex.y = cex.x,
  atcenter = TRUE, add = TRUE, col = NULL, ...)
}
\arguments{
\item{x}{The x coordinates of images to be drawn.}

\item{y}{The y coordinates of images to be drawn.}

\item{obj}{An object of class \code{nativeRaster}.}

\item{file}{A path to either a \code{.png} file or a \code{.jpeg} file.}

\item{cex.x}{A numerical value giving the amount by which the horizontal width of the image should be magnified relative to the default.}

\item{cex.y}{A numerical value giving the amount by which the vertical width of the image should be magnified relative To the default.}

\item{atcenter}{logical. If TRUE x and y coordinates describe the center of the image. Otherwise they represent the bottom-left coordinates of the image.}

\item{add}{logical. Should images be added on the current graph ? If FALSE a new plot is created.}

\item{col}{Optionnal color use to fill pixels whose values are not 0}

\item{...}{Additional arguments to be passed to the \code{rasterImage} function.}
}
\description{
\code{pchimage} returns a plot that displays an image. It enables to users to directly include a png or a jpeg file in a plot region by providing their path.
}
\details{
Either \code{obj} or \code{file} must be defined.

If a \code{file} is defined, \code{readPNG} or \code{readJPEG} according to
the end of the file extension.
}
\examples{
# Example:
img <- png::readPNG(system.file('img', 'Rlogo.png', package='png'), native=TRUE)
n<-15
plot0(c(0,1),c(0,1))
pchImage(0.1+0.8*stats::runif(n), 0.1+0.8*stats::runif(n), cex.x=0.2+1.6*stats::runif(n),
obj=img, angle=360*runif(n), col=2)
}
\keyword{character}
\keyword{image,}
\keyword{plot,}
\keyword{plotting}
