% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pchimage.R
\name{pchImage}
\alias{pchImage}
\title{Images as plotting character}
\usage{
pchImage(
  x,
  y,
  obj = NULL,
  file = NULL,
  cex.x = 1,
  cex.y = cex.x,
  atcenter = TRUE,
  add = TRUE,
  col = NULL,
  ...
)
}
\arguments{
\item{x}{the x coordinates of images to be drawn.}

\item{y}{the y coordinates of images to be drawn.}

\item{obj}{an object of class \code{nativeRaster}.}

\item{file}{a path to either a \code{.png} file or a \code{.jpeg} file.}

\item{cex.x}{a numerical value giving the amount by which the horizontal
width of the image should be magnified, a value of 1 means 5\% of the total width.}

\item{cex.y}{Same as \code{cex.x} for the y axis.}

\item{atcenter}{a logical. If \code{TRUE} them x and y coordinates describe
the center of the image. Otherwise they represent the bottom-left coordinates of the image.}

\item{add}{logical. Should images be added on the current graph? If
\code{FALSE} a new plot is created.}

\item{col}{optional color use to fill pixels whose values are not 0}

\item{...}{Additional arguments to be passed to the \code{rasterImage} function.}
}
\description{
\code{pchimage} returns a plot that displays an image. It enables to users
to directly include a \code{.png} or a \code{.jpeg} file in a plot region.
}
\details{
Either \code{obj} or \code{file} must be defined.
If \code{file} is not null, then \code{\link[png:readPNG]{png::readPNG()}} or \code{\link[jpeg:readJPEG]{jpeg::readJPEG()}} according to
the end of the file extension.
}
\examples{
# Example:
img <- png::readPNG(system.file('img', 'Rlogo.png', package='png'), native=TRUE)
n<-15
plot0(c(0,1),c(0,1))
pchImage(0.1+0.8*stats::runif(n), 0.1+0.8*stats::runif(n), cex.x=0.2+1.6*stats::runif(n),
obj=img, angle=360*runif(n), col=2)
}
\keyword{character}
\keyword{image,}
\keyword{plot,}
\keyword{plotting}
