% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percX.R
\name{percX}
\alias{percX}
\alias{percY}
\title{Values at a given percentage of the axis}
\usage{
percX(percentage = 90)

percY(percentage = 90)
}
\arguments{
\item{percentage}{The percentage of the axis for which the values is returned.}
}
\description{
Return the values of either x-axis or y-axis for a given percentage.
}
\details{
This function intends to ease the positionning of additionnal marks such as text when axis have not common axis.
}
\section{Functions}{
\itemize{
\item \code{percX}: Return the values of x-axis for a given percentage.

\item \code{percY}: Return the values of y-axis for a given percentage.
}}
\examples{
plot0()
text(x=percX(90), y=percY(90), label="cool")
}

