% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pickColors.R
\name{pickcolors}
\alias{pickcolors}
\title{Pick your colors up}
\usage{
pickcolors(ramp = rainbow(1024), nb_shades = 1024, rgb = FALSE)
}
\arguments{
\item{ramp}{A vector of colors used as tone palette.}

\item{nb_shades}{Number of shades to be displayed once a tone is selected.}
}
\value{
A vector
}
\description{
Generate an interactive interface to pick up colors and returns colors selected.
}
\details{
This function generates a graphical window splitted into 6 panels. The top panel serves to select one tone.
The panel right below present \code{nb_shades} of the selected tones.
The bottom rigth panel presents the selected color that can be stored by clicking on the bottom left panel.
The bottom center panel shows the characteristic of the selcted color.
Finally, to abort, the user can simply clik on the left Stop panel
}
\keyword{color,}
\keyword{selection}

