% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pickColors.R
\name{pickColors}
\alias{pickColors}
\title{Pick colors up}
\usage{
pickColors(ramp = rainbow(1024), nb_shades = 1024, rgb = FALSE,
  preview = FALSE)
}
\arguments{
\item{ramp}{A vector of colors used as tone palette.}

\item{nb_shades}{Number of shades to be displayed once a tone is selected.}

\item{rgb}{logical. If TRUE, colors are returned as a rgb matrix. Default is set to FALSE.}

\item{preview}{logical. If TRUE, colors are displayed once the selection is done. Default is set to FALSE.}
}
\value{
A vector
}
\description{
Generate an interactive interface to pick up colors and returns colors selected.
}
\details{
This function generates a graphical window splitted into 6 panels. The top panel serves to select one tone.
The panel right below present \code{nb_shades} of the selected tones.
The bottom rigth panel presents the selected color that can be stored by clicking on the bottom left panel.
The bottom center panel shows the characteristic of the selcted color.
Finally, to abort, the user can simply clik on the left \code{Stop} panel.
}
\keyword{color,}
\keyword{selection}

