% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pickcolors.R
\name{pickcolors}
\alias{pickcolors}
\title{Pick your colors up}
\usage{
pickcolors(ramp = rainbow(1024), colors = 8, tones = colors,
  shades = 32)
}
\arguments{
\item{ramp}{A first set of colors.}

\item{colors}{The number of colors to be selected.}

\item{tones}{A number of tones.}

\item{shades}{A number of shades to be plotted for each tones.}
}
\description{
Generate a simple interactive interface to pick up colors and returns colors selected.
}
\details{
This functions is a two steps color selection. First, color decribed by \code{color} are displayed and users are invited to pick up \code{tones} number of color tones. In the second step, for each selected tones, \code{shades} shades and tints are presented. Users are then allowed to select the final set of \code{colors} colors they require.
}
\note{
Beyond 20 colors, the interest of such function is questionnable.
}
\examples{
#Example 1:
mypal1<-pickcolors()
#Example 2:
mypal2<-pickcolors(colorRampPalette(c('grey','blue','red'))(100), colors=6, tones=4, shades=30)
plot(runif(10),runif(10), col=mypal2, pch=19, cex=3)
}
\keyword{color,}
\keyword{selection}

