% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pickcolors2.R
\name{pickcolors2}
\alias{pickcolors2}
\title{Pick your colors up version 2}
\usage{
pickcolors2(colors = 6, ramp = rainbow(1024))
}
\arguments{
\item{colors}{The number of colors to be selected.}

\item{ramp}{A vector of colors.}
}
\description{
Generate an interactive interface to pick up colors and returns colors selected.
}
\details{
This colors selection open a grapical interface you navigate within following a red frame.
It allows you to confirm at each step wether or not you want to use the tone selected.
It does the same for the shades.
}
\note{
Beyond 20 colors, the interest of such function is questionnable.
}
\keyword{color,}
\keyword{selection}

