% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot0.R
\name{plot0}
\alias{plot0}
\title{An empty plot function}
\usage{
plot0(x = c(-1, 1), y = c(-1, 1), fill = NULL, ...)
}
\arguments{
\item{x}{the x coordinates of points in the plot.}

\item{y}{the y coordinates of points in the plot.}

\item{fill}{The color to be used to fill the plot area.}

\item{...}{additional arguments affecting the plot.}
}
\description{
\code{plot0} returns an empty plot.
}
\details{
The function is based on the \link[graphics]{plot.default} function. It makes the creation of an empty plot faster.
}
\examples{
# Example 1:
plot0()

# Example 2:
plot0(c(-10,10), asp=1)
}
\seealso{
\code{\link[shape]{emptyplot}}
}
\keyword{empty}
\keyword{plot}

