% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotImage.R
\name{plotImage}
\alias{plotImage}
\title{Plot a picture.}
\usage{
plotImage(obj = NULL, file = NULL, add = FALSE, ...)
}
\arguments{
\item{obj}{An object of class \code{nativeRaster} function.}

\item{file}{A path to either a \code{.png} file or a \code{.jpeg} file.}

\item{add}{logical. Should images be added on the current graph? If FALSE a new plot is created.}

\item{...}{Additional arguments to be passed to \code{rasterImage} function.}
}
\description{
Function that returns a plot that displays an image. It enables to users to directly include a png or a jpeg file in a plot region by providing their path.
}
\details{
Either obj or file must be defined.
If a path is provided either \code{readPNG} or \code{readJPEG} according to the end of the file extension.
}
\examples{
# Example:
img <- png::readPNG(system.file("img", "Rlogo.png", package="png"), native=TRUE)
op <- par(no.readonly = TRUE)
par(mfrow=c(4,4), mar=rep(2,4))
for (i in 1:16) plotImage(img)
par(op)
}
\keyword{image.}
\keyword{plot,}

