% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointsInPolygon.R
\name{pointsInPolygon}
\alias{pointsInPolygon}
\title{Are points inside a polygon ?}
\usage{
pointsInPolygon(x, y, poly_x, poly_y)
}
\arguments{
\item{x}{the x coordinates of points to be tested.}

\item{y}{the y coordinates of points to be tested.}

\item{poly_x}{the x coordinates of the static polygons.}

\item{poly_y}{the y coordinates of the static polygons.}
}
\description{
Return a logical vector stating whether or not points are inside a static polygon.
}
\details{
I repoduced the algorithm described in \url{http://alienryderflex.com/polygon/}.
}
\examples{
# Example
plot0(c(0,10), c(0,10))
valx <- 10*runif(20)
valy <- 10*runif(20)
graphics::polygon(c(4,8,8,4),c(4,4,8,8))
graphics::points(valx, valy, col=pointsInPolygon(valx,valy,c(4,8,8,4),c(4,4,8,8))+1)
}
\keyword{over}
\keyword{points,}
\keyword{polygons,}
