% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointsInPolygon.R
\name{pointsInPolygon}
\alias{pointsInPolygon}
\title{Are points inside a polygon?}
\usage{
pointsInPolygon(points, polygon)
}
\arguments{
\item{points}{a matrix of coordinates of points to be tested.}

\item{polygon}{a two-columns matrix including the coordinate of the polygon.}
}
\description{
For a matirx of points coordinates, \code{pointsInPolygon} returns a logical vector stating whether or not these points are inside a specific polygon whose coordinate as passed as an argument.
}
\details{
implements the Ray-casting algorithm
}
\examples{
mat <- matrix(10*runif(100), 50)
res <- pointsInPolygon(mat, cbind(c(4,8,8,4),c(4,4,8,8)))
# Visual assessment
plot0(c(0,10), c(0,10))
graphics::polygon(c(4,8,8,4),c(4,4,8,8))
graphics::points(mat[,1], mat[,2], col=res+1)
}
\references{
\url{https://rosettacode.org/wiki/Ray-casting_algorithm#C}
}
\keyword{over}
\keyword{points,}
\keyword{polygons,}
