% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polarPlot.R
\name{polarPlot}
\alias{polarPlot}
\title{Polar Plot}
\usage{
polarPlot(seqtime, seqval = NULL, rad = 1, from = 0, to = 2 * pi,
  incr = 0.005, labelc = NULL, tckcol = 1, atc = NULL, labelr = NULL,
  atr = NULL, clockwise = TRUE, n_signif = 2, add = FALSE, ...)
}
\arguments{
\item{seqtime}{sequence of time values (equivalent to the x axis).}

\item{seqval}{sequence of values of interest (radial axis, equivalent to the y axis).}

\item{rad}{radius of the circle.}

\item{from}{starting point of the circle.}

\item{to}{ending point of the circle.}

\item{incr}{increment used to draw the circle.}

\item{labelc}{character or expression vector of labels to be placed at the tickpoints.}

\item{tckcol}{color of the tickmarks.}

\item{atc}{the points at which tick-marks are to be drawn. By default (when \code{NULL}) tickmark locations are computed.}

\item{labelr}{character or expression vector specifying the _text_ to be placed at the radial-axis marks.}

\item{atr}{the points at which radial-axis marks are to be drawn.}

\item{clockwise}{logical. If TRUE, the plot must de read clockwise, otherwise, counter-clockwise.}

\item{n_signif}{number of significant numbers to be displayed (used when labelc is \code{NULL}).}

\item{add}{logical. add to current plot?}

\item{...}{additional argument to be passed to \code{polygon} function.}
}
\description{
Draw a polar plot.
}
\details{
Polar Plot
}
\examples{
polarPlot(1:40, stats::runif(40), to=1.9*pi, col='grey30', border='grey80')
}
\seealso{
\code{\link[plotrix]{polar.plot}}
}
