% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showPalette.R
\name{showPalette}
\alias{showPalette}
\title{Show a color Palette}
\usage{
showPalette(ramp, inline = FALSE, add_number = FALSE, cex_num = 1.2)
}
\arguments{
\item{ramp}{A vector of colors.}

\item{inline}{logical. If TRUE, the colors are displayed on a single row.}

\item{add_number}{logical. If TRUE, color vector's indices are added.}

\item{cex_num}{The maginification coefficient of the color vector's indices.}
}
\description{
Generate an interactive interface to pick up colors and returns colors selected.
}
\details{
This function generates a graphical window splitted into 6 panels. The top panel serves to select one tone.
}
\examples{
showPalette(floor(runif(30,1,9)), add_number=TRUE)
}
\keyword{color,}
\keyword{selection}
