% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation.R
\name{translation}
\alias{translation}
\title{Translation}
\usage{
translation(x, y, xtrans = 0, ytrans = 0, add = FALSE, ...)
}
\arguments{
\item{x}{the x coordinates of points. It can also be a matrix (see details).}

\item{y}{the y coordinates of points.}

\item{xtrans}{the x corrdinate of the translation vector.}

\item{ytrans}{the y corrdinate of the translation vector.}

\item{add}{logical. If \code{TRUE} the set of transforned points is drawn as a polygon.}

\item{...}{Additionnal arguments to be passed to \code{polygon} function (used only if \code{add} is TRUE).}
}
\description{
Compute a translation for a set of points. The transformed set of points is
optionnaly add to the current plot.
}
\details{
For details about what is a translation, see \url{https://en.wikipedia.org/wiki/Translation_(geometry)}.
Note that if \code{x} is a matrix with more than 2 columns, then x is the
first column and y the second one.
}
\examples{
# Example 1:
plot0(c(0,10),c(0,10))
x <- c(4,6,5)
y <- c(2,2,4)
graphics::polygon(x,y)
trans1 <- translation(x,y,xtrans=2,ytrans=5, add=TRUE, border=4, lwd=2)

# Example 2:
x <- c(2,4,3,1)
y <- c(1,1,3,3)
plot0(c(0,10),c(0,10))
graphics::polygon(x,y)
for (i in 1:6) translation(x,y,xtrans=i,ytrans=i, add=TRUE, border=i+1, lwd=2)
}
\keyword{geometry}
\keyword{translation,}
